Fixed Modal Issues
A few issues arose with adding in a modal. The first was it not
ending the game once activated, but this was fixed by
clearing the interval. The second was being able to close
the modal by clicking outside of it, but was fixing using a
small line added to the callModal function.